Add files via upload
Definition, Types and Factors A portfolio’s meaning can be defined as a collection of financial assets and investment tools that are held by an individual, a financial institution or an investment firm. To develop a profitable portfolio, it is essential to become familiar with its fundamentals and the factors that influence it. What is a Portfolio?